서버에 다녀온 응답이 성공일 경우, ui 표현해야할 실제 데이터를 뿌려줌(score는 float로 변경해야 나타남)

+++ app/src/main/java/com/neppplus/self_apipractice_20220106/models/ReviewData.kt
@@ -6,7 +6,7 @@ class ReviewData(
     val id : Int,
     val title : String,
     val content : String,
-    val score : Double,
+    val score : Float,
     val thumbnail_img : String,
     val user : UserData,
 ) : Serializable {
